@@ -1,12 +1,10 @@
-const {
+import {
   removeAtag,
   getAndRemoveConfig,
   getAndRemoveDocisfyIgnorConfig,
-} = require('../../src/core/render/utils');
-
-const { tree } = require(`../../src/core/render/tpl`);
-
-const { slugify } = require(`../../src/core/render/slugify`);
+} from '../../src/core/render/utils.js';
+import { tree } from '../../src/core/render/tpl.js';
+import { slugify } from '../../src/core/render/slugify.js';
 
 // Suite
 // -----------------------------------------------------------------------------
