@@ -1,4 +1,4 @@
-const prettierConfig = require('./.prettierrc');
+const prettierConfig = require('./.prettierrc.json');
 
 module.exports = {
   root: true,
