@@ -2,7 +2,7 @@
  * Fork https://github.com/egoist/docute/blob/master/src/utils/front-matter.js
  */
 /* eslint-disable */
-import parser from './yaml'
+import parser from './yaml.js'
 
 var optionalByteOrderMark = '\\ufeff?'
 var pattern =
