@@ -1,11 +1,7 @@
-//
+import { jest } from '@jest/globals';
 import { greet } from './fixtures/greet.js';
 import { getTimeOfDay } from './fixtures/get-time-of-day.js';
 import * as getTimeOfDayModule from './fixtures/get-time-of-day.js';
-
-// const greet = require('./fixtures/greet');
-// const getTimeOfDay = require('./fixtures/get-time-of-day');
-// const getTimeOfDayModule = { getTimeOfDay: getTimeOfDay };
 
 // Suite
 // -----------------------------------------------------------------------------
@@ -53,20 +49,26 @@ describe(`Example Tests`, function () {
     });
   });
 
+  // Test not working, but it is an example so no matter.
+  // "TypeError: setSystemTime is not available when not using modern timers"
   describe('Fake Timers', function () {
-    // jest version issue
-    // test('data & time', () => {
-    //   jest.useFakeTimers();
-    //   jest.setSystemTime(fakeDate);
-    //   const timeOfDay = getTimeOfDay();
-    //   expect(timeOfDay).toContain;
-    // });
+    test.skip('data & time', () => {
+      const fakeDate = new Date().setHours(1);
+
+      jest.useFakeTimers('modern');
+      jest.setSystemTime(fakeDate);
+
+      const timeOfDay = getTimeOfDay();
+
+      expect(timeOfDay).toBe('morning');
+    });
   });
 
-  describe('Mocks & Spies', function () {
-    test('mock import/require dependency using jest.fn()', () => {
-      const testModule = require('./fixtures/get-time-of-day.js');
-      const { greet: testGreet } = require('./fixtures/greet.js');
+  describe('Mocks & Spys', function () {
+    // TODO use jest.unstable_mockModule(), experimental support for mocking ES Modules, https://github.com/facebook/jest/issues/10025
+    test.skip('mock import dependency using jest.fn()', async () => {
+      const testModule = { ...(await import('./fixtures/get-time-of-day.js')) };
+      const { greet: testGreet } = { ...(await import('./fixtures/greet.js')) };
 
       testModule.getTimeOfDay = jest.fn(() => 'day');
 
@@ -77,7 +79,8 @@ describe(`Example Tests`, function () {
       expect(greeting).toBe(`Good day, John!`);
     });
 
-    test('mock import/require dependency using jest.doMock()', () => {
+    // TODO use jest.unstable_mockModule(), experimental support for mocking ES Modules, https://github.com/facebook/jest/issues/10025
+    test.skip('mock import dependency using jest.doMock()', async () => {
       const mockModulePath = './fixtures/get-time-of-day.js';
 
       jest.doMock(mockModulePath, () => ({
@@ -85,8 +88,8 @@ describe(`Example Tests`, function () {
         getTimeOfDay: jest.fn(() => 'night'),
       }));
 
-      const mockGetTimeOfDay = require(mockModulePath).getTimeOfDay;
-      const { greet: testGreet } = require('./fixtures/greet.js');
+      const mockGetTimeOfDay = (await import(mockModulePath)).getTimeOfDay;
+      const { greet: testGreet } = await import('./fixtures/greet.js');
 
       const timeOfDay = mockGetTimeOfDay();
       const greeting = testGreet('John');
@@ -104,7 +107,8 @@ describe(`Example Tests`, function () {
       expect(Math.random()).toBe(0.1);
     });
 
-    test('spy on import/require dependency using jest.spyOn()', () => {
+    // TODO use jest.unstable_mockModule(), experimental support for mocking ES Modules, https://github.com/facebook/jest/issues/10025
+    test.skip('spy on import dependency using jest.spyOn()', () => {
       jest
         .spyOn(getTimeOfDayModule, 'getTimeOfDay')
         .mockImplementation(() => 'night');
