@@ -1,4 +1,4 @@
-import { getTimeOfDay } from './get-time-of-day';
+import { getTimeOfDay } from './get-time-of-day.js';
 
 export function greet(name = 'friend') {
   const timeOfDay = getTimeOfDay();
